Implement dropzeros and droptol using fkeep.

+++ csparse.h
@@ -17,7 +17,7 @@
 #include <string>
 #include <vector>
 
-typedef std::size_t csint;  // TODO just change to "int" for simplicity
+typedef std::uint64_t csint;  // TODO just change to "int" for simplicity?
 
 #include "csc.h"  // must be first for conversion
 #include "coo.h"
